Refactor security functions to use bcrypt directly, update database connection defaults, enhance history model with Float type for latency, integrate MLflow tracking in RAG, add CI/CD pipeline, and implement tests for API and endpoints.

+++ app/models/history.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey,Float
 from datetime import datetime
 from app.database.database import Base
 
@@ -10,5 +10,5 @@ class AnswersHistory(Base):
     user_id = Column(Integer, ForeignKey("users.id"))
     answer = Column(String)
     question = Column(String)
-    latency_ms = Column(Integer)
+    latency_ms = Column(Float)
     cluster = Column(Integer)
